--- generic-validator/reqirements-osi-3/osi_trafficsign.txt
@@ -1,24 +1,0 @@
-# Example Requirements table for OSI Traffic Sign 
-
-# TrafficSignValue
-
-TrafficSignValue.value is-set
-TrafficSignValue.value_unit is-set
-
-TrafficSignValue.Unit is-defined
-
-# TrafficSign
-TrafficSign.id 
-TrafficSign.id is-valid
-TrafficSign.id is-minimum 0
-TrafficSign.id is-maximum 9999999999
-
-TrafficSign.main_sign is-valid
-TrafficSign.supplementary_sign is-valid 
-
-TrafficSign.MainSign.base is-valid
-TrafficSign.MainSign.classification is-valid
-
-# Main Sign Classification
-TrafficSign.MainSign.Classification.variability  is-set
-
